Refactor RabbitMQ-related components and add refund event handling
- Consolidate RabbitMQ-related packages under `com.nexaworks.rafiq.rabbit`.
- Add `RefundRequestEvent` and `RefundSucceededEvent` records for event handling.
- Introduce `RabbitConfiguration` for message converter setup.
- Update constants and imports to align with the new structure.
