+++ src/main/java/com/nexaworks/rafiq/dto/event/refund/RefundSucceededEvent.java
@@ -1,0 +1,7 @@
+package com.nexaworks.rafiq.dto.event.refund;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record RefundSucceededEvent(UUID refundId, Instant occurredAt) {
+}
