+++ src/main/java/com/nexaworks/rafiq/rabbit/event/RefundRequestEvent.java
@@ -1,0 +1,7 @@
+package com.nexaworks.rafiq.rabbit.event;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record RefundRequestEvent(UUID consultationId, Instant occurredAt) {
+}
