+++ src/main/java/com/nexaworks/rafiq/rabbit/notificaiton/SmsNotification.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.notificaiton;
+package com.nexaworks.rafiq.rabbit.notificaiton;
 
 public record SmsNotification() implements Notification {
 }
