+++ src/main/java/com/nexaworks/rafiq/rabbit/publisher/SmsNotificationPublisher.java
@@ -1,11 +1,9 @@
 package com.nexaworks.rafiq.rabbit.publisher;
 
-import static com.nexaworks.rafiq.constant.RabbitMQConstant.*;
-
 import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.stereotype.Component;
 
-import com.nexaworks.rafiq.dto.notificaiton.SmsNotification;
+import com.nexaworks.rafiq.rabbit.notificaiton.SmsNotification;
 
 import lombok.extern.slf4j.Slf4j;
