+++ src/main/java/com/nexaworks/rafiq/service/notification/SmsNotificationService.java
@@ -3,7 +3,7 @@ package com.nexaworks.rafiq.service.notification;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-import com.nexaworks.rafiq.dto.notificaiton.SmsNotification;
+import com.nexaworks.rafiq.rabbit.notificaiton.SmsNotification;
 
 import lombok.extern.slf4j.Slf4j;
